% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListSubscriptionsResponse}
\alias{ListSubscriptionsResponse}
\title{ListSubscriptionsResponse Object}
\usage{
ListSubscriptionsResponse(subscriptions = NULL, nextPageToken = NULL)
}
\arguments{
\item{subscriptions}{The subscriptions that match the request}

\item{nextPageToken}{If not empty, indicates that there may be more subscriptions that match the request; this value should be passed in a new `ListSubscriptionsRequest` to get more subscriptions}
}
\value{
ListSubscriptionsResponse object
}
\description{
ListSubscriptionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListSubscriptions` method.
}

