% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{projects.subscriptions.acknowledge}
\alias{projects.subscriptions.acknowledge}
\title{Acknowledges the messages associated with the `ack_ids` in the`AcknowledgeRequest`. The Pub/Sub system can remove the relevant messagesfrom the subscription.Acknowledging a message whose ack deadline has expired may succeed,but such a message may be redelivered later. Acknowledging a message morethan once will not result in an error.}
\usage{
projects.subscriptions.acknowledge(AcknowledgeRequest, subscription)
}
\arguments{
\item{AcknowledgeRequest}{The \link{AcknowledgeRequest} object to pass to this method}

\item{subscription}{The subscription whose message is being acknowledged}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other AcknowledgeRequest functions: \code{\link{AcknowledgeRequest}}
}
