% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{projects.subscriptions.modifyAckDeadline}
\alias{projects.subscriptions.modifyAckDeadline}
\title{Modifies the ack deadline for a specific message. This method is usefulto indicate that more time is needed to process a message by thesubscriber, or to make the message available for redelivery if theprocessing was interrupted. Note that this does not modify thesubscription-level `ackDeadlineSeconds` used for subsequent messages.}
\usage{
projects.subscriptions.modifyAckDeadline(ModifyAckDeadlineRequest, subscription)
}
\arguments{
\item{ModifyAckDeadlineRequest}{The \link{ModifyAckDeadlineRequest} object to pass to this method}

\item{subscription}{The name of the subscription}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other ModifyAckDeadlineRequest functions: \code{\link{ModifyAckDeadlineRequest}}
}
