% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{projects.subscriptions.modifyPushConfig}
\alias{projects.subscriptions.modifyPushConfig}
\title{Modifies the `PushConfig` for a specified subscription.This may be used to change a push subscription to a pull one (signified byan empty `PushConfig`) or vice versa, or change the endpoint URL and otherattributes of a push subscription. Messages will accumulate for deliverycontinuously through the call regardless of changes to the `PushConfig`.}
\usage{
projects.subscriptions.modifyPushConfig(ModifyPushConfigRequest, subscription)
}
\arguments{
\item{ModifyPushConfigRequest}{The \link{ModifyPushConfigRequest} object to pass to this method}

\item{subscription}{The name of the subscription}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other ModifyPushConfigRequest functions: \code{\link{ModifyPushConfigRequest}}
}
