% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{projects.topics.delete}
\alias{projects.topics.delete}
\title{Deletes the topic with the given name. Returns `NOT_FOUND` if the topicdoes not exist. After a topic is deleted, a new topic may be created withthe same name; this is an entirely new topic with none of the oldconfiguration or subscriptions. Existing subscriptions to this topic arenot deleted, but their `topic` field is set to `_deleted-topic_`.}
\usage{
projects.topics.delete(topic)
}
\arguments{
\item{topic}{Name of the topic to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}
}
