% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PubsubMessage}
\alias{PubsubMessage}
\title{PubsubMessage Object}
\usage{
PubsubMessage(PubsubMessage.attributes = NULL, data = NULL,
  attributes = NULL, messageId = NULL, publishTime = NULL)
}
\arguments{
\item{PubsubMessage.attributes}{The \link{PubsubMessage.attributes} object or list of objects}

\item{data}{The message payload}

\item{attributes}{Optional attributes for this message}

\item{messageId}{ID of this message, assigned by the server when the message is published}

\item{publishTime}{The time at which the message was published, populated by the server when it receives the `Publish` call}
}
\value{
PubsubMessage object
}
\description{
PubsubMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message data and its attributes. The message payload must not be empty; it must contain either a non-empty data field, or at least one attribute.
}
\seealso{
Other PubsubMessage functions: \code{\link{PubsubMessage.attributes}}
}

