% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PushConfig.attributes}
\alias{PushConfig.attributes}
\title{PushConfig.attributes Object}
\usage{
PushConfig.attributes()
}
\value{
PushConfig.attributes object
}
\description{
PushConfig.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Endpoint configuration attributes.Every endpoint has a set of API supported attributes that can be used tocontrol different aspects of the message delivery.The currently supported attribute is `x-goog-version`, which you canuse to change the format of the push message. This attributeindicates the version of the data expected by the endpoint. Thiscontrols the shape of the envelope (i.e. its fields and metadata).The endpoint version is based on the version of the Pub/SubAPI.If not present during the `CreateSubscription` call, it will default tothe version of the API used to make such call. If not present during a`ModifyPushConfig` call, its value will not be changed. `GetSubscription`calls will always return a valid version, even if the subscription wascreated without this attribute.The possible values for this attribute are:* `v1beta1`: uses the push format defined in the v1beta1 Pub/Sub API.* `v1` or `v1beta2`: uses the push format defined in the v1 Pub/Sub API.
}
\seealso{
Other PushConfig functions: \code{\link{PushConfig}}
}
