% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ReceivedMessage}
\alias{ReceivedMessage}
\title{ReceivedMessage Object}
\usage{
ReceivedMessage(ackId = NULL, message = NULL)
}
\arguments{
\item{ackId}{This ID can be used to acknowledge the received message}

\item{message}{The message}
}
\value{
ReceivedMessage object
}
\description{
ReceivedMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message and its corresponding acknowledgment ID.
}

