% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(pushConfig = NULL, ackDeadlineSeconds = NULL, name = NULL,
  topic = NULL)
}
\arguments{
\item{pushConfig}{If push delivery is used with this subscription, this field is}

\item{ackDeadlineSeconds}{This value is the maximum time after a subscriber receives a message}

\item{name}{The name of the subscription}

\item{topic}{The name of the topic from which this subscription is receiving messages}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subscription resource.
}
\seealso{
Other Subscription functions: \code{\link{projects.subscriptions.create}}
}
